Mise en place des roles, revue des formulaires avec ajout, update et suppression OK

+++ src/Controller/TeamController.php
@@ -5,15 +5,19 @@ namespace App\Controller;
 use App\Entity\Player;
 use App\Entity\Team;
 use App\Form\AddPlayerToTeamType;
+use App\Form\PlayerType;
 use App\Form\TeamType;
 use App\Security\EmailVerifier;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-use Symfony\Component\Security\Core\User\User;
 use Symfony\Component\Security\Csrf\TokenGenerator\TokenGeneratorInterface;
 
 
@@ -55,25 +59,113 @@ class TeamController extends AbstractController
         $teamForm->handleRequest($request);
 
         if ($teamForm->isSubmitted() && $teamForm->isValid()) {
-//            $teamName = $team->getName();
-//            $testTeam = $em->getRepository(Team::class)->findByName($teamName);
-//            if ($testTeam->getId() == $team->getId()){
-//                $testTeam = null;
-//            }
-//            if ($testTeam != null) {
-//                $this->addFlash("alert-danger", "This Team name is already used");
-//            } else {
-            $team->addPlayer($player);
-            $em->persist($team);
-            $em->flush();
-            $this->addFlash("success", "Welcome to your Team !");
-            return $this->redirectToRoute('teamDetail', array('id' => $team->getId()));
-//            }
+            $teamName = $request->get('name');
+            $testTeam = $em->getRepository(Team::class)->findByName($teamName);
+
+            if ($testTeam != null) {
+                $this->addFlash("warning", "This Team name is already used");
+            } else {
+                $photo = $teamForm->get('logo')->getData();
+                if ($photo) {
+                    $originalFilename = pathinfo($photo->getClientOriginalName(), PATHINFO_FILENAME);
+                    // On inclut le nom du fichier à l'URL
+                    $safeFilename = transliterator_transliterate('Any-Latin; Latin-ASCII; [^A-Za-z0-9_] remove; Lower()', $originalFilename);
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $photo->guessExtension();
+
+                    // On enregistre le fichier dans le dossier demandé
+                    try {
+                        $photo->move(
+                            $this->getParameter('logoTeam_directory'),
+                            $newFilename
+                        );
+                    } catch (FileException $e) {
+
+                    }
+
+                    $team->setLogo($newFilename);
+                }
+                $team->setName($teamName);
+                $team->addPlayer($player);
+                $player->setRoles(['ROLE_TEAM_ADMIN']);
+                $em->persist($team);
+                $em->flush();
+                $this->addFlash("success", "Welcome to your Team !");
+                return $this->redirectToRoute('teamDetail', array('id' => $team->getId()));
+            }
 
         }
         return $this->render('team/addTeam.html.twig', [
             'teamForm' => $teamForm->createView(),
 
+        ]);
+    }
+
+    /**
+     * @Route("/updateTeam/{id}", name="updateTeam")
+     * @param Request $request
+     * @param EntityManagerInterface $entityManager
+     * @return RedirectResponse|Response
+     */
+    public function updateTeam(Request $request, EntityManagerInterface $entityManager, $id)
+    {
+        if (!$this->getUser()) {
+            $this->addFlash('danger', "You cannot access to this page if you are not connected");
+            return $this->redirectToRoute('main');
+        }
+//        $this->denyAccessUnlessGranted('ROLE_TEAM_ADMIN');
+
+        $team = $entityManager->getRepository(Team::class)->find($id);
+
+        $teamForm = $this->createForm(TeamType::class, $team);
+        $teamForm->handleRequest($request);
+
+        if ($teamForm->isSubmitted() && $teamForm->isValid()) {
+
+            $photo = $teamForm->get('logo')->getData();
+            if ($photo) {
+                $originalFilename = pathinfo($photo->getClientOriginalName(), PATHINFO_FILENAME);
+                // On inclut le nom du fichier à l'URL
+                $safeFilename = transliterator_transliterate('Any-Latin; Latin-ASCII; [^A-Za-z0-9_] remove; Lower()', $originalFilename);
+                $newFilename = $safeFilename . '-' . uniqid() . '.' . $photo->guessExtension();
+
+                // On enregistre le fichier dans le dossier demandé
+                try {
+                    $photo->move(
+                        $this->getParameter('logoTeam_directory'),
+                        $newFilename
+                    );
+                } catch (FileException $e) {
+
+                }
+
+                $team->setLogo($newFilename);
+            }
+
+            $testName = $team->getName();
+            $testTeam = $entityManager->getRepository(Team::class)->findOneBy(
+                ['name' => $testName]
+            );
+
+            if ($testTeam->getId() == $team->getId()) {
+                $testTeam = null;
+            }
+
+            if ($testTeam != null) {
+                $this->addFlash("alert-danger", "This name is already used");
+            } else {
+                $team->setName($testName);
+                $entityManager->flush();
+                $this->addFlash("success", "Team updated !");
+
+                return $this->redirectToRoute('teamDetail', ['id' => $team->getId()]);
+
+            }
+
+        }
+
+        return $this->render('team/updateTeam.html.twig', [
+            'teamForm' => $teamForm->createView(),
+            'team' => $team
         ]);
     }
 
@@ -88,6 +180,7 @@ class TeamController extends AbstractController
         }
 
         $team = $em->getRepository(Team::class)->find($id);
+        $player = $this->getUser();
         $players = $team->getPlayers();
 
         $addPlayerToTeamForm = $this->createForm(AddPlayerToTeamType::class);
@@ -97,31 +190,35 @@ class TeamController extends AbstractController
             $selectedPlayer = $request->get('selectNicknames');
             $playerAddedToTeam = $em->getRepository(Player::class)->findOneBy(['nickname' => $selectedPlayer]);
 
-            $token = $tokenGenerator->generateToken();
-            try {
-                $playerAddedToTeam->setResetToken($token);
-                $em->flush();
-            } catch (\Exception $e) {
-                $this->addFlash('warning', $e->getMessage());
-                return $this->redirectToRoute('teamDetail', array('id' => $team->getId()));
+            if (!in_array($playerAddedToTeam, $players->toArray())) {
+                $token = $tokenGenerator->generateToken();
+                try {
+                    $playerAddedToTeam->setResetToken($token);
+                    $em->flush();
+                } catch (\Exception $e) {
+                    $this->addFlash('warning', $e->getMessage());
+                    return $this->redirectToRoute('teamDetail', array('id' => $team->getId()));
+                }
+                $url = $this->generateUrl('playerConfirmJoinTeam', array('token' => $token, 'id' => $id), UrlGeneratorInterface::ABSOLUTE_URL);
+
+                $message = (new \Swift_Message('A Team wants to recruit you !'))
+                    ->setFrom('teamRecruitment@esportevents.com')
+                    ->setTo($playerAddedToTeam->getEmail())
+                    ->setBody(
+                        $this->renderView(
+                            'emails/teamRecruitment.html.twig',
+                            ['playerAddedToTeam' => $playerAddedToTeam,
+                                'team' => $team,
+                                'url' => $url]
+                        ),
+                        'text/html'
+                    );
+
+                $mailer->send($message);
+                $this->addFlash("success", "The player has been asked to join your Team");
+            } else {
+                $this->addFlash("warning", "The player is already in your Team");
             }
-            $url = $this->generateUrl('playerConfirmJoinTeam', array('token' => $token, 'id' => $id), UrlGeneratorInterface::ABSOLUTE_URL);
-
-            $message = (new \Swift_Message('A Team wants to recruit you !'))
-                ->setFrom('teamRecruitment@esportevents.com')
-                ->setTo($playerAddedToTeam->getEmail())
-                ->setBody(
-                    $this->renderView(
-                        'emails/teamRecruitment.html.twig',
-                        ['playerAddedToTeam' => $playerAddedToTeam,
-                            'team' => $team,
-                            'url' => $url]
-                    ),
-                    'text/html'
-                );
-
-            $mailer->send($message);
-            $this->addFlash("success", "The player has been asked to join your Team");
             return $this->redirectToRoute('teamDetail', array('id' => $team->getId()));
 
         }
@@ -130,6 +227,7 @@ class TeamController extends AbstractController
             [
                 "team" => $team,
                 'players' => $players,
+                'player' => $player,
                 'addPlayerToTeamForm' => $addPlayerToTeamForm->createView()
 
             ]
@@ -159,8 +257,6 @@ class TeamController extends AbstractController
             'team' => $team,
             'id' => $id
         ]);
-
-
     }
 
     /**
@@ -168,14 +264,13 @@ class TeamController extends AbstractController
      */
     public function searchPlayer(EntityManagerInterface $em, Request $request)
     {
-        $search = $request->get('add_player_to_team_nickname');
+        $search = $request->get('search');
         $searchPlayer = $em->getRepository(Player::class)->searchPlayer($search);
         return $this->json($searchPlayer, 200, [], ['groups' => 'group1']);
     }
 
 
     /**
-     *
      * @Route("/deleteTeam/{id}", name="deleteTeam")
      */
     public function deleteTeam(EntityManagerInterface $em, $id)
@@ -184,6 +279,8 @@ class TeamController extends AbstractController
             $this->addFlash('danger', "You cannot access to this page if you're not connected");
             return $this->redirectToRoute('main');
         }
+        $this->denyAccessUnlessGranted('ROLE_TEAM_ADMIN');
+
         $team = $em->getRepository(Team::class)->find($id);
 
         $em->remove($team);
@@ -195,6 +292,34 @@ class TeamController extends AbstractController
         return $this->render("team/index.html.twig",
             [
                 "teams" => $teams
+            ]
+        );
+    }
+
+    /**
+     * @Route("/removePlayerFromTeam/{id}", name="removePlayerFromTeam")
+     */
+    public function removePlayerFromTeam(EntityManagerInterface $em, $id)
+    {
+        if (!$this->getUser()) {
+            $this->addFlash('danger', "You cannot access to this page if you're not connected");
+            return $this->redirectToRoute('main');
+        }
+
+        $player = $em->getRepository(Player::class)->find($id);
+        $team = $player->getTeam();
+        $team->removePlayer($player);
+        $players = $team->getPlayers();
+        $em->persist($team);
+        $em->flush();
+        $this->addFlash("success", "The player has been removed from your Team");
+
+        return $this->render("team/teamDetail.html.twig",
+            [
+                'player' => $player,
+                'players' => $players,
+                'id' => $id,
+                'team' => $team
             ]
         );
     }
